+++ src/ntlMenu.js
@@ -1,0 +1,8 @@
+
+let ntlMenu = {
+	execute: function(args) {
+		const ntl = require('../node_modules/ntl/cli.js');
+	}
+};
+
+module.exports = ntlMenu;