+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 ## package-menu-cli Changelog
 
 ### Overview
-See README.md for installation and configuration details
+See `README.md` for installation and configuration details
 
 ### Just the changes
+
+- 1.4.0: Added `launch` interface
+    - Further cli options will be passed to `launcher`
+      - `$ package-menu -l -a | -m | -i`
+          `  --all, -a run all scripts`
+          `  --info, -i shows package description`
+          `  --multiple, -m allow multiple selection`
 
 - 1.3.0: Added `compress` switch
     - `$ package-menu --compress`
@@ -11,7 +18,7 @@ See README.md for installation and configuration details
 
 - 1.1.1: Minor tweaks
     - Removed fixed column width for 'script' column
-    
+
 - 1.1.0: Internal refactoring for better testing
 
 - 1.0.1: Adjust table width